% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scen_baseline}
\alias{scen_baseline}
\title{Specification of the baseline simulation scenarios (2.104 iterations)}
\format{
A data frame with 2.104 rows and 12 variables. :
\describe{
   \item{n}{int. number of subjects}
   \item{n_t}{int. number of observed data points for each subject}
   \item{zeta}{float. between-subject variance for the means of the emission distributions}
   \item{Q}{float. between-subject variance for the intercepts of the transition-probability matrix}
   \item{scenario_id}{string .unique id of the simulation scenario}
   \item{rank}{int. iteration number of the simulation scenario}
   \item{iteration_id}{string. unique id of the iteration}
   \item{dsim_seed}{int. random seed used to generate the data}
   \item{model_seed}{int. random seed used to run the mHMM}
   \item{start_gamma}{json. initial values for the between-subject transition probability matrix}
   \item{start_emiss}{json. intiial values for each of the 3 emission distributions}
}
}
\usage{
scen_baseline
}
\description{
This dataset contains simulation settings for 2.104 baseline iterations  of my simulation study.
It was created using the function \link[sleepsimR]{generate_scenarios} with seed value 912326.
}
\seealso{
Versions 1.5.1 <https://github.com/JasperHG90/sleepsimR-api/releases/tag/v1.5.1> and
1.5.2 <https://github.com/JasperHG90/sleepsimR-api/releases/tag/v1.5.2> of the
sleepsimR-api program contain this dataset to manage the simulation study. For more information,
visit <https://github.com/JasperHG90/sleepsimR-api>
}
\keyword{datasets}
