% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scen_out_zeta_025}
\alias{scen_out_zeta_025}
\title{Simulation results for the baseline scenarios, zeta=0.25 (scenarios 1-5A)}
\format{
A data frame with 27 rows and 32 variables. :
\describe{
   \item{emiss_var_short}{string. One of 'EEG_mean_beta', 'EOG_median_theta' or 'EOG_min_beta'.}
   \item{state}{string. Latent state. One of 'state1' (Awake), 'state2' (NREM) or 'state3' (REM) or transitions from state to state.}
   \item{pbias_scenario_{1-5}}{String. Percent bias of the parameter estimates versus the population parameter for a scenario. MC error given in parentheses.}
   \item{empirical_se_scenario_{1-5}}{String. Empirical SE for a scenario. MC error given in parentheses.}
   \item{modSE_scenario_{1-5}}{String. Average model SE/CCI for a scenario. MC error given in parentheses.}
   \item{MSE_scenario_{1-5}}{String. MSE for a scenario. MC error given in parentheses.}
   \item{coverage_scenario_{1-5}}{String. Percent coverage for a scenario. MC error given in parentheses.}
   \item{bias_corr_coverage_scenario_{1-5}}{numeric. Percent bias-corrected coverage for a scenario. MC error given in parentheses.}
}

A data frame with 27 rows and 32 variables. :
\describe{
   \item{emiss_var_short}{string. One of 'EEG_mean_beta', 'EOG_median_theta' or 'EOG_min_beta'.}
   \item{state}{string. Latent state. One of 'state1' (Awake), 'state2' (NREM) or 'state3' (REM) or transitions from state to state.}
   \item{pbias_scenario_{1-5}}{String. Percent bias of the parameter estimates versus the population parameter for a scenario. MC error given in parentheses.}
   \item{empirical_se_scenario_{1-5}}{String. Empirical SE for a scenario. MC error given in parentheses.}
   \item{modSE_scenario_{1-5}}{String. Average model SE/CCI for a scenario. MC error given in parentheses.}
   \item{MSE_scenario_{1-5}}{String. MSE for a scenario. MC error given in parentheses.}
   \item{coverage_scenario_{1-5}}{String. Percent coverage for a scenario. MC error given in parentheses.}
   \item{bias_corr_coverage_scenario_{1-5}}{numeric. Percent bias-corrected coverage for a scenario. MC error given in parentheses.}
}
}
\usage{
scen_out_zeta_025

scen_out_zeta_025
}
\description{
This dataset contains the simulation results for baseline scenarios 1-5A. To reproduce these datasets, follow the instructions in the
"data-raw/5_preprocess_baseline_results" found in this R library.

This dataset contains the simulation results for baseline scenarios 1-5B. To reproduce these datasets, follow the instructions in the
"data-raw/5_preprocess_baseline_results" found in this R library.
}
\seealso{
The simulation was conducted using the following two programs. The first program is the resource
manager found at <https://github.com/JasperHG90/sleepsimR-api/releases>. In particular, versions 1.5.1 - 1.5.4
were used. These versions contain different baseline scenarios. For more
information, see \link[sleepsimRdata]{scen_baseline}. The program used to run the simulations can be found here
<https://github.com/JasperHG90/sleepsimR-run/releases>. In particular, versions 1.5.1 - 1.5.2 were used. The
simulations were executed on a cluster managed by SURF <https://www.surf.nl/en>. For more information
about the architectural design of the simulation study, visit
<https://github.com/JasperHG90/sleepsimR-documentation>.
For a definition of the simulation metrics, see: Morris, Tim P., Ian R. White, and Michael J. Crowther. "Using simulation studies to evaluate statistical methods." Statistics in medicine 38.11 (2019): 2074-2102.

The simulation was conducted using the following two programs. The first program is the resource
manager found at <https://github.com/JasperHG90/sleepsimR-api/releases>. In particular, versions 1.5.1 - 1.5.4
were used. These versions contain different baseline scenarios. For more
information, see \link[sleepsimRdata]{scen_baseline}. The program used to run the simulations can be found here
<https://github.com/JasperHG90/sleepsimR-run/releases>. In particular, versions 1.5.1 - 1.5.2 were used. The
simulations were executed on a cluster managed by SURF <https://www.surf.nl/en>. For more information
about the architectural design of the simulation study, visit
<https://github.com/JasperHG90/sleepsimR-documentation>.
For a definition of the simulation metrics, see: Morris, Tim P., Ian R. White, and Michael J. Crowther. "Using simulation studies to evaluate statistical methods." Statistics in medicine 38.11 (2019): 2074-2102.
}
\keyword{datasets}
