% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scen_rerun}
\alias{scen_rerun}
\title{Specification of the simulation scenarios (862 iterations)}
\format{
A data frame with 6.912 rows and 12 variables. :
\describe{
   \item{n}{int. number of subjects}
   \item{n_t}{int. number of observed data points for each subject}
   \item{zeta}{float. between-subject variance for the means of the emission distributions}
   \item{Q}{float. between-subject variance for the intercepts of the transition-probability matrix}
   \item{scenario_id}{string .unique id of the simulation scenario}
   \item{rank}{int. iteration number of the simulation scenario}
   \item{iteration_id}{string. unique id of the iteration}
   \item{dsim_seed}{int. random seed used to generate the data}
   \item{model_seed}{int. random seed used to run the mHMM}
   \item{start_gamma}{json. initial values for the between-subject transition probability matrix}
   \item{start_emiss}{json. intiial values for each of the 3 emission distributions}
}
}
\usage{
scen_rerun
}
\description{
This dataset contains simulation settings for 862 iterations (approx. 6 iterations
per simulation scenario) of my simulation study. This dataset contains instructions to execute
a second MCMC chain for specific models. The models from the original simulations are then compared
to the second chain to evaluate whether or not they have converged to their stable posterior
distributions. This dataset is an adaption of the dataset found in \link[sleepsimRdata]{subs} and
was created using the R file found in "data-raw/2_make_simulation_scenarios/make_simulation_scenarios.R"
}
\seealso{
Version 1.3.3 <https://github.com/JasperHG90/sleepsimR-api/releases/tag/v1.3.3> of the
sleepsimR-api program contains this dataset to manage the simulation study. For more information,
visit <https://github.com/JasperHG90/sleepsimR-api>
}
\keyword{datasets}
