% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation_data_emiss_varmu}
\alias{simulation_data_emiss_varmu}
\title{Parsed simulation results for emission distribution between-subject variances.}
\format{
A list with two entries:
data_preprocessed: A data frame with 323.937 rows and 14 variables. :
\describe{
   \item{iteration_id}{string. unique id of the iteration}
   \item{emiss_var_short}{string. One of 'EEG_mean_beta', 'EOG_median_theta' or 'EOG_min_beta'.}
   \item{state}{string. Latent state. One of 'state1' (Awake), 'state2' (NREM) or 'state3' (REM)}
   \item{median}{numeric. MAP value (median) of the posterior distribution.}
   \item{SE}{numeric. MAP value (SE) of the posterior distribution.}
   \item{var}{string. Concatenation of 'emiss_var_short' plus the statistic (median/SE) plus the
   state name.}
   \item{true_val}{numeric. Value of the population parameter used in the simulation study. These values are
   the same as those used in the sleepsimR library. See
   <https://github.com/JasperHG90/sleepsimR/blob/master/R/zzz.R>}
   \item{lower}{numeric. Lower 95\% CCI for the median MAP value.}
   \item{upper}{numeric. Upper 95\% CCI for the median MAP value.}
   \item{scenario_id}{string .unique id of the simulation scenario}
   \item{n}{int. number of subjects}
   \item{n_t}{int. number of observed data points for each subject}
   \item{zeta}{float. between-subject variance for the means of the emission distributions}
   \item{Q}{float. between-subject variance for the intercepts of the transition-probability matrix}
}
summary_by_scenario: A data frame with 1.296 rows and 21 variables. In my study, I use 144 scenarios. Given
that I use three emission variables, each of which has 3 latent states, each scenario is represented 9 times
in this dataset.
\describe{
   \item{scenario_id}{string .unique id of the simulation scenario}
   \item{emiss_var_short}{string. One of 'EEG_mean_beta', 'EOG_median_theta' or 'EOG_min_beta'.}
   \item{bias}{bias of the parameter estimates in the scenario relative to the population parameter.}
   \item{bias_mcmc_se}{MCMC standard error of bias estimate.}
   \item{empirical_se}{Empirical standard error of the parameter estimates in the scenario.}
   \item{empirical_se_mcmc_se}{MCMC standard error of the empirical standard error.}
   \item{MSE}{MSE of the parameter estimates in the scenario relative to the population parameter.}
   \item{MSE_mcmc_se}{MCMC standard error of the MSE estimate.}
   \item{coverage}{Percentage of cases in a scenario where the population parameter is included in
   the 95\% CCI.}
   \item{coverage_mcmc_se}{MCMC standard error of the coverage estimate.}
   \item{bias_corr_coverage}{Percentage of cases in a scenario where the average estimated parameter
   value is included in the 95\% CCI.}
   \item{bias_corr_coverage_mcmc_se}{MCMC standard error of the bias corrected coverage estimate.}
   \item{multimodal}{P-value of a test for multimodality in the distribution of parameter estimates.
   See \link[multimode]{modetest}.}
   \item{modSE}{Model standard error of the parameter estimates in the scenario.}
   \item{modSE_mcmc_se}{MCMC standard error of the modSE estimate.}
   \item{n}{int. number of subjects}
   \item{n_t}{int. number of observed data points for each subject}
   \item{zeta}{float. between-subject variance for the means of the emission distributions}
   \item{Q}{float. between-subject variance for the intercepts of the transition-probability matrix}
}
}
\usage{
simulation_data_emiss_varmu
}
\description{
This dataset contains two separate datasets in a list. The first dataset is raw data that contains all
parameter estimates obtained from the simulation results (median MAP, SE MAP, lower 95\% CCI and upper
95\% CCI). The second dataset summarizes the raw in terms of the simulation metrics (bias, MSE etc.) for
each simulation scenario. To reproduce these datasets, follow the instructions in the
"data-raw/3_preprocess_simulation_results" found in this R library.
}
\seealso{
The simulation was conducted using the following two programs. The first program is the resource
manager found at <https://github.com/JasperHG90/sleepsimR-api/releases>. In particular, versions 1.3.1,
1.3.2 and 1.3.3 were used. These versions contain different iterations of the simulation study. For more
information, see \link[sleepsimRdata]{scen}, \link[sleepsimRdata]{scen_subs} and
\link[sleepsimRdata]{scen_rerun}. The program used to run the simulations can be found here
<https://github.com/JasperHG90/sleepsimR-run/releases>. In particular, version 1.3 was used. The
simulations were executed on a cluster managed by SURF <https://www.surf.nl/en>. For more information
about the architectural design of the simulation study, visit
<https://github.com/JasperHG90/sleepsimR-documentation>.
For a definition of the simulation metrics, see: Morris, Tim P., Ian R. White, and Michael J. Crowther. "Using simulation studies to evaluate statistical methods." Statistics in medicine 38.11 (2019): 2074-2102.
}
\keyword{datasets}
