% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sleepdata}
\alias{sleepdata}
\title{Preprocessed EEG and EOG channels that I use in my thesis.}
\format{
A data frame with 58947 rows and 6 variables. Each subject has 1.440 occasions
(=time points), except subject 20, who has 1.347 occasions.:
\describe{
  \item{id}{Unique id of the subject}
  \item{sleep_state}{Label of the state at occasion t}
  \item{EEG_Fpz_Cz_mean_theta}{Value of the logit-transformed EEG channel at occasion t}
  \item{EEG_Fpz_Cz_mean_beta}{Value of the logit-transformed EEG channel at occasion ts}
  \item{EOG_median_theta}{Value of the logit-transformed EOG channel at occasion t}
  \item{EOG_min_beta}{Value of the logit-transformed EOG channel at occasion t}
}
}
\source{
\url{https://physionet.org/content/sleep-edfx/1.0.0/}
}
\usage{
sleepdata
}
\description{
This dataset contains 4 EEG and EOG channels that I use to conduct my
simulation study and empirical application.
}
\seealso{
The docker app used to download the data can be found here:
<>. The docker app used to preprocess the data can be found here:
<>. The script used generate this dataset can be found here:
<>.
}
\keyword{datasets}
