% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{summary_statistics_sleepdata}
\alias{summary_statistics_sleepdata}
\title{Summary statistics collected for the 'sleepdata' dataset.}
\format{
A data frame with 12 rows and 4 variables:
\describe{
  \item{variable}{Name of the emission distribution variable.}
  \item{sleep_state}{Label of the latent sleep state.}
  \item{mmvar}{Mean of the emission distribution.}
  \item{vvvar}{Between-subject variance of the mean of the emission distribution.}
}
}
\source{
\url{https://physionet.org/content/sleep-edfx/1.0.0/}
}
\usage{
summary_statistics_sleepdata
}
\description{
This dataset contains summary statistics for the \link[sleepsimRdata]{sleepdata} dataset.
}
\keyword{datasets}
