% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{total_variance_sleepdata}
\alias{total_variance_sleepdata}
\title{Total variance found in each emission distribution.}
\format{
A data frame with 12 rows and 3 variables:
\describe{
  \item{variable}{Name of the emission distribution variable.}
  \item{sleep_state}{Label of the latent sleep state.}
  \item{tvar}{Total variance for each emission distribution.}
}
}
\source{
\url{https://physionet.org/content/sleep-edfx/1.0.0/}
}
\usage{
total_variance_sleepdata
}
\description{
This dataset contains the total variance for the emission distributions found in
the \link[sleepsimRdata]{sleepdata} dataset.
}
\keyword{datasets}
